Move beanvalidation tests to standalone junits
