+++ etc/jenkins/pr_verify.groovy
@@ -202,6 +202,7 @@ spec:
                     sh """
                                 mvn -B -V clean install -pl :eclipselink
                                 mvn -B -V verify -pl :org.eclipse.persistence.jpa.modelgen.processor,:org.eclipse.persistence.jpa.jse.test,:org.eclipse.persistence.extension,:org.eclipse.persistence.jpa.jpql,:org.eclipse.persistence.jpa.wdf.test,:org.eclipse.persistence.jpars,:org.eclipse.persistence.dbws,:org.eclipse.persistence.dbws.builder,:eclipselink,:org.eclipse.persistence.distribution.tests -P mysql;
+                                mvn -B -V -P mysql -amd verify -pl :org.eclipse.persistence.jpa.testapps;
                             """
                 }
             }
