+++ jpa/eclipselink.jpa.test/pom.xml
@@ -359,8 +359,6 @@
                                 <item>eclipselink-remote</item>
                                 <item>jpa-performance</item>
                                 <item>jpa-performance2</item>
-                                <item>eclipselink-beanvalidation-model</item>
-                                <item>eclipselink-beanvalidation-dynamic-model</item>
                                 <item>eclipselink-identity-model</item>
                                 <item>eclipselink-jpa22-model</item>
                                 <item>eclipselink-xml-extended-composite-advanced-model</item>
@@ -498,8 +496,6 @@
                                 <item>eclipselink-metamodel-model</item>
                                 <item>eclipselink-delimited-model</item>
                                 <item>eclipselink-extensibility-model</item>
-                                <item>eclipselink-beanvalidation-model</item>
-                                <item>eclipselink-beanvalidation-dynamic-model</item>
                                 <item>eclipselink-xml-mapping-metadata-complete</item>
                                 <item>eclipselink-xml-composite-advanced-model-member_1</item>
                                 <item>eclipselink-xml-extended-model</item>
@@ -748,12 +744,6 @@
                                     <name>server-test-jpa-xmltest</name>
                                     <properties>
                                         <puName>default</puName>
-                                    </properties>
-                                </itemWithProperty>
-                                <itemWithProperty>
-                                    <name>server-test-jpa-beanvalidation</name>
-                                    <properties>
-                                        <puName>beanvalidation</puName>
                                     </properties>
                                 </itemWithProperty>
                                 <itemWithProperty>
@@ -1280,8 +1270,6 @@
                                 <additionalClasspathElement>${project.build.directory}/${wave.file.prefix}eclipselink-metamodel-model.jar</additionalClasspathElement>
                                 <additionalClasspathElement>${project.build.directory}/${wave.file.prefix}eclipselink-delimited-model.jar</additionalClasspathElement>
                                 <additionalClasspathElement>${project.build.directory}/${wave.file.prefix}eclipselink-extensibility-model.jar</additionalClasspathElement>
-                                <additionalClasspathElement>${project.build.directory}/${wave.file.prefix}eclipselink-beanvalidation-model.jar</additionalClasspathElement>
-                                <additionalClasspathElement>${project.build.directory}/${wave.file.prefix}eclipselink-beanvalidation-dynamic-model.jar</additionalClasspathElement>
                                 <additionalClasspathElement>${project.build.directory}/eclipselink-composite-advanced-model.jar</additionalClasspathElement>
                                 <additionalClasspathElement>${project.build.directory}/${wave.file.prefix}eclipselink-composite-advanced-model-member_1.jar</additionalClasspathElement>
                                 <additionalClasspathElement>${project.build.directory}/${wave.file.prefix}eclipselink-composite-advanced-model-member_2.jar</additionalClasspathElement>
@@ -1642,20 +1630,6 @@
                 <extra-model>eclipselink-jpa22-model-server</extra-model>
                 <packageName>jpa22</packageName>
                 <test-client-class>org.eclipse.persistence.testing.tests.jpa22.advanced.JPA22TestSuite</test-client-class>
-            </properties>
-        </profile>
-        <profile>
-            <id>server-test-jpa-beanvalidation</id>
-            <properties>
-                <skip-test-jpa-srg>true</skip-test-jpa-srg>
-                <skip-server-test-jpa>false</skip-server-test-jpa>
-                <skip-jpa-jse-build>true</skip-jpa-jse-build>
-                <puName>beanvalidation</puName>
-                <!--modelname property name is used there due substitution string in application.xml-->
-                <modelname>eclipselink-beanvalidation-model</modelname>
-                <extra-model>eclipselink-beanvalidation-model</extra-model>
-                <packageName>beanvalidation</packageName>
-                <test-client-class>org.eclipse.persistence.testing.tests.jpa.beanvalidation.BeanValidationJunitTest</test-client-class>
             </properties>
         </profile>
         <profile>
