--- jpa/eclipselink.jpa.test/src/it/java/org/eclipse/persistence/testing/models/jpa/beanvalidation/Address.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2009, 2020 Oracle and/or its affiliates. All rights reserved.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0,
- * or the Eclipse Distribution License v. 1.0 which is available at
- * http://www.eclipse.org/org/documents/edl-v10.php.
- *
- * SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
- */
-
-// Contributors:
-
-package org.eclipse.persistence.testing.models.jpa.beanvalidation;
-
-import jakarta.persistence.Embeddable;
-import jakarta.validation.constraints.Size;
-
-@Embeddable
-public class Address {
-    String street;
-    String city;
-    @Size(max = 2)
-    String state;
-
-    public Address() {}
-
-    public Address(String street, String city, String state) {
-         this.street = street;
-         this.city = city;
-         this.state = state;
-    }
-}
