--- jpa/eclipselink.jpa.test/src/it/java/org/eclipse/persistence/testing/models/jpa/beanvalidation/Project.java
@@ -1,45 +1,0 @@
-/*
- * Copyright (c) 2009, 2020 Oracle and/or its affiliates. All rights reserved.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0,
- * or the Eclipse Distribution License v. 1.0 which is available at
- * http://www.eclipse.org/org/documents/edl-v10.php.
- *
- * SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
- */
-
-// Contributors:
-
-package org.eclipse.persistence.testing.models.jpa.beanvalidation;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.validation.constraints.Size;
-
-@Entity(name="CMP3_BV_PROJECT")
-public class Project {
-    public static final int NAME_MAX_SIZSE = 5;
-
-    @Id
-    private int        id;
-    @Size(max = NAME_MAX_SIZSE)
-    private String    name;
-
-    public Project() {}
-
-    public Project(int id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-}
