--- jpa/eclipselink.jpa.test/src/it/resources/eclipselink-beanvalidation-dynamic-model/employee-orm.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2018, 2020 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0 which is available at
-    http://www.eclipse.org/legal/epl-2.0,
-    or the Eclipse Distribution License v. 1.0 which is available at
-    http://www.eclipse.org/org/documents/edl-v10.php.
-
-    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
-
--->
-
-<entity-mappings version="2.1"
-	xmlns="http://www.eclipse.org/eclipselink/xsds/persistence/orm"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-
-	<package>org.eclipse.persistence.testing.models.jpa.beanvalidation</package>
-
-	<entity class="DynamicEmployee" access="VIRTUAL">
-		<attributes>
-			<id name="id" attribute-type="Integer">
-				<generated-value strategy="SEQUENCE" generator="employee-generator" />
-				<sequence-generator name="employee-generator" sequence-name="DYN_EMP_SEQ" />
-			</id>
-			<basic name="name" attribute-type="String"/>
-			<basic name="surname" attribute-type="String"/>
-			<version name="version" attribute-type="Long"/>
-		</attributes>
-	</entity>
-
-</entity-mappings>
