--- jpa/eclipselink.jpa.test/src/it/resources/eclipselink-beanvalidation-dynamic-model/server/employee-constraints.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2018, 2020 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0 which is available at
-    http://www.eclipse.org/legal/epl-2.0,
-    or the Eclipse Distribution License v. 1.0 which is available at
-    http://www.eclipse.org/org/documents/edl-v10.php.
-
-    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
-
--->
-
-<constraint-mappings xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                     xsi:schemaLocation="http://jboss.org/xml/ns/javax/validation/mapping validation-mapping-1.0.xsd"
-                     xmlns="http://jboss.org/xml/ns/javax/validation/mapping">
-    <default-package>org.eclipse.persistence.testing.models.jpa.beanvalidation</default-package>
-    <bean class="DynamicEmployee" ignore-annotations="true">
-        <getter name="name" ignore-annotations="true">
-            <constraint annotation="jakarta.validation.constraints.NotNull" />
-        </getter>
-        <getter name="surname" ignore-annotations="true">
-            <constraint annotation="jakarta.validation.constraints.NotNull" />
-        </getter>
-    </bean>
-</constraint-mappings>
