+++ jpa/eclipselink.jpa.testapps/jpa.test.beanvalidation.dynamic/src/main/resources-server/persistence.xml
@@ -1,0 +1,32 @@
+<!--
+
+    Copyright (c) 2018, 2022 Oracle and/or its affiliates. All rights reserved.
+
+    This program and the accompanying materials are made available under the
+    terms of the Eclipse Public License v. 2.0 which is available at
+    http://www.eclipse.org/legal/epl-2.0,
+    or the Eclipse Distribution License v. 1.0 which is available at
+    http://www.eclipse.org/org/documents/edl-v10.php.
+
+    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
+
+-->
+
+<persistence xmlns="http://java.sun.com/xml/ns/persistence" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/persistence persistence_2_0.xsd" version="2.0">
+    <persistence-unit name="@default@" transaction-type="@transaction-type@">
+        <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
+        <@datasource-type@>@data-source-name@</@datasource-type@>
+        <mapping-file>META-INF/employee-orm.xml</mapping-file>
+        <exclude-unlisted-classes>true</exclude-unlisted-classes>
+        <validation-mode>CALLBACK</validation-mode>
+        <properties>
+            <property name="eclipselink.target-server" value="@server-platform@"/>
+            <property name="eclipselink.target-database" value="@database-platform@"/>
+            <property name="eclipselink.weaving" value="static"/>
+            <property name="eclipselink.validate-existence" value="true"/>
+            <property name="eclipselink.logging.logger" value="DefaultLogger"/>
+            <property name="eclipselink.jdbc.allow-native-sql-queries" value="true"/>
+            <property name="jakarta.persistence.schema-generation.database.action" value="create"/>
+        </properties>
+    </persistence-unit>
+</persistence>
