+++ jpa/eclipselink.jpa.testapps/jpa.test.beanvalidation/src/main/resources-server/persistence.xml
@@ -1,0 +1,35 @@
+<!--
+
+    Copyright (c) 2018, 2022 Oracle and/or its affiliates. All rights reserved.
+
+    This program and the accompanying materials are made available under the
+    terms of the Eclipse Public License v. 2.0 which is available at
+    http://www.eclipse.org/legal/epl-2.0,
+    or the Eclipse Distribution License v. 1.0 which is available at
+    http://www.eclipse.org/org/documents/edl-v10.php.
+
+    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
+
+-->
+
+<persistence xmlns="https://jakarta.ee/xml/ns/persistence"
+             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+             xsi:schemaLocation="https://jakarta.ee/xml/ns/persistence https://jakarta.ee/xml/ns/persistence/persistence_3_0.xsd"
+             version="3.0">
+    <persistence-unit name="@default@" transaction-type="@transaction-type@">
+        <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
+        <@datasource-type@>@data-source-name@</@datasource-type@>
+        <class>org.eclipse.persistence.testing.models.jpa.beanvalidation.Employee</class>
+        <class>org.eclipse.persistence.testing.models.jpa.beanvalidation.Project</class>
+        <class>org.eclipse.persistence.testing.models.jpa.beanvalidation.Address</class>
+        <validation-mode>CALLBACK</validation-mode>
+        <properties>
+            <property name="eclipselink.target-server" value="@server-platform@"/>
+            <property name="eclipselink.target-database" value="@database-platform@"/>
+            <property name="eclipselink.weaving" value="@server-weaving@"/>
+            <property name="eclipselink.validate-existence" value="true"/>
+            <property name="eclipselink.logging.logger" value="DefaultLogger"/>
+            <property name="eclipselink.jdbc.allow-native-sql-queries" value="true"/>
+        </properties>
+    </persistence-unit>
+</persistence>
