+++ pom.xml
@@ -1968,7 +1968,7 @@
                         <plugin>
                             <groupId>org.jacoco</groupId>
                             <artifactId>jacoco-maven-plugin</artifactId>
-                            <version>0.8.6</version>
+                            <version>0.8.8</version>
                         </plugin>
                     </plugins>
                 </pluginManagement>
@@ -1976,6 +1976,11 @@
                     <plugin>
                         <groupId>org.jacoco</groupId>
                         <artifactId>jacoco-maven-plugin</artifactId>
+                        <configuration>
+                            <append>true</append>
+                            <dumpOnExit>true</dumpOnExit>
+                            <classDumpDir>${project.build.outputDirectory}</classDumpDir>
+                        </configuration>
                         <executions>
                             <execution>
                                 <id>jacoco-prepare-agent</id>
@@ -2032,6 +2037,9 @@
                 <module>utils/eclipselink.dbws.builder.test.oracle.server</module>
 
                 <module>utils/eclipselink.utils.sigcompare</module>
+
+                <!-- testapps -->
+                <module>jpa/eclipselink.jpa.testapps</module>
             </modules>
         </profile>
         <profile>
